Get s3270 going again after the big x3270 ICU removal.
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@159 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/Common/glue.c
@@ -1124,7 +1124,7 @@ read_resource_file(const char *filename, Boolean fatal)
 		/* If this line is a continuation, try again. */
 		if (bsl) {
 			ilen += strlen(buf + ilen);
-			if (ilen >= sizeof(buf) - 1) {
+			if ((unsigned)ilen >= sizeof(buf) - 1) {
 				(void) sprintf(where, "%s:%d: Line too long\n",
 				    filename, lno);
 				Warning(where);
